requirements.txt
#python==3.10
python-dotenv==1.0.1
openai==0.28
streamlit==1.29.0
Pillow==10.1.0
requests==2.31.0